pyproject.toml
[tool.poetry]
name = "project-transforge"
version = "0.1.0"
description = ""
authors = ["Daniil Gnidenko <dagnidenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pandas = "^2.3.2"
openpyxl = "^3.1.5"
loguru = "^0.7.3"
requests = "^2.32.5"
types-requests = "^2.32.4.20250809"
pandas-stubs = "^2.3.2.250827"
python-dotenv = "^1.1.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
mypy = "^1.17.1"
black = "^25.1.0"
isort = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '''
/(\.git)/
'''

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'